refactor: consolidate interpretation tests
